feat: use oclif/test v4 (#1420)
* test: use oclif/test v4

* chore: clean up

* test: use new major of oclif/test

* chore: bump commitlint

* feat: generate CLI with oclif/test v4

* test: dont use createSandbox

+++ src/readme-generator.ts
@@ -251,7 +251,7 @@ USAGE
   }
 
   protected async write(file: string, content: string): Promise<void> {
-    await fs.outputFile(file, content)
+    if (!this.options.dryRun) await fs.outputFile(file, content)
   }
 
   /**
